github-token-manager/requirements.txt
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
PyJWT==2.10.1
requests==2.32.5
